Changed visual speed round handling, added conflict notification.
Changed the way the header displays speed round status. Instead of
changing the header background color, now only the text of the speed
round button turns red.

When the announced score does not match the calculated score, the
background of the announced score area turns red. This may be changed to
make it more obvious later by changing the whole header and/or add
another visual notification.