Unified noFX version into master using command line argument.
Instead of parallel branches, TriviaClient now decides whether to run
with JavaFX based on a command line argument. The default is not to use
JavaFX, but if the first argument is useFX, the client will create the
split pane and use JavaFX to show the IRC web client.

The NoJavaFX branch is now abandoned.