Overrode default enter behavior for text areas.
Added a function to TriviaDialogPanel that adds key bindings to a
JTextArea that map Enter to clicking OK on the option pane and
Shift-Enter to a carriage return. (Issue #4)

Made the new question text area on the confirm question overwrite
editable. (Issue #25)

Renamed QuestionEntryWindow to OpenQuestionDialog for consistency with
other classes.

Added CTRL+O command on the workflow tab to open an Open Question
Dialog.