Implemented saving of sort method and hide question state.
The hide question state is now saved to the settings file to persist
between sessions. (Issue #43)
The current sort method is now saved to the settings file to persist
between sessions. (Issue #44)
Fixed bug where changing the sort method by clicking on the header did
not change the selected sort in the menu.