Fixed bug with some windows not remembering their position. (Issue #53)
Fixed bug with some windows not remembering their position. This
occurred on windows that were setting a custom name. Since the position
was being loaded as the TriviaDialog was created, changing the name
afterwards was too late for the changed name to impact the setting being
loaded, but on close the settings were saved under the new name. Moved
the loadPosition call to the WindowOpened event so it occurs after the
new name is set.