First wave of add/close tab changes. (Issue #71)
Added tracking of all available tabbed panes to handle ChangeListeners
that hide context menus.
Added tracking of frames to handle WindowFocusListeners that hide
context menus.
Created a Hashtable in TriviaClient to hold all of the data panels.
Added methods to retrieve a list of available panels and to get a named
panel.