Major revamp to client framework improving structure and providing more flexibility.
Moved all of the code for creating the first window to TriviaFrame
(formerly FloatingPanel). This unifies the code for individual frames
and now all windows will have menus and status bars. The program exits
when all windows are closed.

Created a Properties object in TriviaClient to hold settings. This
allows for additional customization later by changing static variables
to properties.

Saved properties file is now read only once at startup instead of every
time a setting is needed. Properties file is only saved when program
ends.

Changed most of the static variables in TriviaClient to properties.