Added auto-parsing of visual trivia to add links. (Issue #46)
Question text areas were turned into custom text panes that
automatically replace references to visual trivia to hyperlinks to the
appropriate visual trivia page. This is done using the following regular
expression for flexibility in input formatting: "([Vv]isual )([Tt]rivia
)?(#)?([0-9])+".

Renamed TriviaPanel to TriviaMainPanel, then extracted most of the class
to a new TriviaPanel that serves as a common superclass to both
TriviaMainPanel and TriviaDialogPanel.