Added graphic for add tab tab.
Added a green plus icon for the tab that adds a tab.

Made the add tab tab disabled so it can't be selected as a tab, but
overrode the disabled tab painter so it still looks like the same.

Fixed bug with the QuestionPane flickering when there are scroll bars.

Fixed bug with first window getting an iterated title (e.g., "Trivia
(1)")