Revamped some functions to better handle opening of questions. (Issue #98)
Changed the way questions are opened to operate better when questions
are answered before typing in the question has been completed. Open
function now only takes the question number, and automatically sets the
question text to "<user> is typing in the question". When the question
information is entered, the question text and value are set directly
rather than re-calling open().

Additional methods added to TriviaServer, Trivia Interface, Trivia, and
Round to set question text and value directly.