Added audio when a new answer is added to the queue. (Issue #105)
Added the JLayer library to provide support for playing MP3 files.

Added functions to TriviaClient to play a sound.

Added check in AnswerQueuePanel to see if the length of the queue has
changed and have the client play a sound if it has.

Volume control to be added later. (Issue #123)