Added functionality to change the parent frame when a tab is moved to a new window. (Issue #124)
Added functionality to change the parent frame when a tab is moved to a
new window. As part of this change, client and frame were made part of
TriviaMainPanel (nearly all of the children had both of these, so it
really should've been in the superclass anyway). Added method to
TriviaMainPanel to change the parent frame. Overrode that method in
compound panels (Workflow and History) to also call changeFrame on all
the child panels.

Refactored some methods to make the (client, frame) order more
consistent.