Changing the status of an answer now works correctly if the answer queue updates and changes the location. (Issue #107)
Changing the status of an answer now changes the answer associated with
the combo box when it was opened, even if that line changes location
because the queue updates. This is accomplished by adding a variable
activeComboBox to the AnswerQueueSubPanel that is set when a comboBox is
opened. When an ItemChanged event is received from a comboBox, the name
of this.activeComboBox is used (if not set, the name of the source is
used).

This seems a little kludgy, but was the issue was caused by the way Java
handles open combo boxes rather than a bug. If you check the name of a
combo box that has moved, you get the name of the box that is in that
position when the item is selected, even if it is not the box that was
clicked.