Added boolean isLive to AnswerQueuePanel.
Added a boolean to AnswerQueuePanel that is set to whether the queue
should be live or not. If it is not live, the round is set by the
constructor and can be changed, and is not updated to the current round
number when the queue is updated. This is to facilitate future work to
allow review of queues for ended rounds. (Issue #122)