Conversion to JSON. (Issue #110, Issue #111) Converted to MAVEN for dependency management.
Massive overhaul to switch from RMI to Websockets & JSON.

Removal of all RMI methods as well as both Interfaces. RMI callback
didn't work when tested beyond a local system.

Creation of ClientMessage and ServerMessage to hold data send between
from client to server and from server to client respectively. Each has
an enumeration Command that specifies what the payload of the data is
and what action should be taken. Each also contains an Encoder and
Decoder that uses jackson to convert the data object to/from a JSON.
Finally, they contain a Factory with a method for each command type to
generate a message based on the appropriate arguments.

Renamed TriviaServer to TriviaServerEndpoint. This name makes more sense
because the whole-server data (most notably the main Trivia object) were
converted into static quantities. Many other methods were made static as
well.

Both TriviaClient and TriviaServerEndpoint received a number of methods
to accomodate the message passing. Both have the Endpoint methods
onOpen, onMessage, onError and onClose. onMessage then has a case
statement that takes the command within the received message and takes
the appropriate action for the message type. Both were also given static
sendMessage methods that take the appropriate Message type and handle
the try/catch block for sending the message. This makes it simple to
send messages from other methods.

All RMI calls were converted to websocket calls using sendMessage and
the MessageFactory. This made the code quite a bit cleaner since it
eliminated the former pattern of try/catch that counted failed RMI calls
and aborted if there were too many fails. The new paradigm requires only
a single line.

Removal of the UserList class. With the websocket connections, the
tracking of users is easier and now moved into the TriviaServerEndpoint
class.

Each TriviaServerEndpoint instantiation tracks the username, role, last
active time, timeToIdle, and last Round versions sent for each session
as a client connects. When a command is received that changes trivia
data, the changes are automatically pushed to all connected clients.
Likewise when there is a new user or a role change, updated user lists
are pushed. Probably need to also add a timer to refresh the user lists
anyway every so often to show update idleness. (Issue #124) When a
connection is closed, the associated user is automatically removed from
the list.

The specialized data structures (Trivia, Round, Question, Answer,
ScoreEntry) were augmented with annotations to ensure they are
reconstructed properly in the JSON encode/decode process when passes
between server and client.