Added timer to update idle users. (Issue #126) Fixed bug where reopening a question did not mark it as incorrect.
Added a timer to ServerEndpoint to rebuild user lists to determine idle
users when there are otherwise no changes to the user lists.

Added a reopen method to Question, Round and Trivia to reopen a question
and mark it as incorrect. Changed onMessage in ServerEndpoint to use the
new function.