Fixed bug where answer agreements would carry over to new rounds. (Issue #108)
Added a reset method to AnswerQueueSubPanel that is called by
AnswerQueuePanel when the round advances for a live panel. Agreement is
still only enforced a client level, so the state of the button will not
carry over if another instance of the GUI is launched (but a +1 will
still be added into the total, allowing a user to agree multiple times.

This is still not the most satisfying solution, but a more sophisticated
treatment server-side might not be worth the trouble, as it would
require not only altering the agree/disagree methods all the way from
ServerEndpoint to Answer (simple), but also the addition of methods to
change user names in all Answers when changed.