Audited code to move sending of commands off of the event thread. Changed the way opening of questions is handled.
Went through all of the client code for calls to sendMessage that would
run on the event dispatch thread and moved them to run on a SwingWorker
thread. This should improve responsiveness (and is appropriate
practice).

Questions are now only "opened" once, when the button is first pressed.
Finishing the opening of the question (by submitting the question text
and value) are now done with an edit operation. This should make the
code flow clearer.