Added disagree button (Issue #109), agreement now tracked by-user (Issue #108). (Also adresses Issue #137, Issue #139)
Added a disagree (-1) button, rearranging the confidence/agreement
region in the answer queue. The disagree button is in a custom
ButtonGroup with agree button so that only one can be pressed at a time,
but also allows for unselecting both.

Changed model for tracking agreement to a Hashtable in Answer that
relates user names to agree/neutral/disagree, so agreement is now
tracked on a per-user basis instead of per GUI instance. Replaced
agree/disagree methods with a changeAgreement method that propagates
from Answer through TriviaServerEndpoint and back to TriviaClientMessage
and AnswerQueuePanel. Clicking agree/disagree will now persist if the
user logs in again, and will propagate to other instance of the client
with the same user name.

Propagation of user name changes will be implemented in a future commit.