Added a method to restart the server timer. (Issue #165)
Added a method to restart the timer in TriviaServerEndpoint and set up
the save/standings/user schedules. This is a failsafe in case the timer
stops due to an exception.

Added a ClientMessage command to trigger the timer restart. Added an
item to the Admin menu that sends the command.