User name changes now update the user names in Q&As for current round. (Issue #159)
Added changeName method to Answer & Question to update the name of
caller, submitter and the agreement hash.

Added changeName method to Round that changes the name for all questions
& answers in the round. Trivia changeName calls changeName on the
current round. When TriviaServerEndpoint receives a change name command,
it now calls the method on the Trivia to change the name in the current
round's Q&As.