Made changes to immediately register closed/correct questions (Issue #176). Removed operator from questions (only on Answers now).
Changed the flow when a question is closed or marked as correct to send
a notice to the server immediately to close the question, with the
operator or answer text added at a later time. This mirrors the process
when questions are opened. (Issue #176)

Making this change required the addition of several new message commands
to accommodate the second wave of information, as well as changes in all
of the close and markCorrect methods and additional methods to handle
the second wave.

Also took this opportunity to remove the operator from the Question
object. Operator remains a field in Answer objects, and I intend to add
an operator dialog for other call-in results. This will make our record
keeping more thorough for question challenges (as often as we don't get
points for a question we got correct is being told an answer was
incorrect that was correct and we need to track who those operators
are).