Added operator dialogs for call-in responses other than correct.
Changed the name of the CorrectEntryDialog to OperatorDialog to reflect
its new function. The title for the dialog is now specified in the
constructor.

Added an OperatorDialog follow each of the other call-in results
(Incorrect, Partially Correct). This will allow us to track the
interactions with the phone bank better for question challenging
purposes.