Added variable for number of visual trivia and way to set.
Added a variable to Trivia for the number of visual trivia. This is in
partial fulfillment of Issue #186. Added getters/setters for this
variable.

Added an item to the Admin menu to set the number of visual trivia.
Added a new dialog to handle input of same. Added new client message
type and associated variable, getters, and factory method to pass the
number of visual trivia to server. Added handlers to TriviaServer to
process the new message.

Added functions to broadcast the number of teams and the number of
visual trivia to all clients.

Renamed broadcastTrivia to broadcastChangedRounds to more accurately
describe function.