Added context menu to edit the show and host names.
Added popup menus to the show name and host name labels that pop-up text
fields to set the show and host names. (Issue #184)

Fixed client message to set the round number when passing the show name
or host name.

Changed SummaryPanel to use the action name instead of the source to
determine the requested action for all actions performed.