Converted answer text filter and discrepancy to use popup text field.
Changed the discrepancy text popup menu to now contain a text field to
edit the discrepancy text without a dialog.

Changed the answer filter popup menu to have a submenu for the text
filter with a text field instead of spawning a dialog.

Changed the queue answer filter menu to have a submenu for the text
filter with a text field instead of spawning a dialog.

Since menus cannot have have a key accelerator like a menu item can,
replaced the accelerator that spawned the filter text dialog with an
anonymous KeyEventDispatcher that intercepts the shortcut and creates a
text field at the current mouse location. This is done with a popup menu
that contains a second qFilterTextField.

Removed the classes for the discrepancy and answer filter dialogs.

Changed lower bound on nVisualTriviaSpinner to 0 instead of 1 to prevent
exception when nVisualTrivia hasn't been set yet.