Made default confidence —. Rearranged confidence/agreement panel.
Changed new answer dialog to have a 0 confidence unless the slider is
moved. 0 confidence is displayed as — on the slider and in the answer
queue. (Issue #161)

Rearranged the confidence/agreement panel in the answer queue. (Issue
#160)

Fixed intermittent NPE when loading if the answer filter text fields try
to set before the pattern is set in TriviaGUI.