Moved disposal of wait dialog to the event dispatch thread.
The wait dialog is now disposed of using invokeLater so it runs on the
event dispatch thread. Hopefully, this solves Issue #178.

Added a check to updateGUI of TriviaFrame to make sure the we don't try
to set the text in the answer text filter text fields before they have
been created. This was done by adding a field initComplete that is set
to true at the end of object creation. updateGUI will sleep until that
is complete.