Changed user list to use User objects. Fixed bug with user activity.
Changed the list model in UserListPanel to be a list of User instead of
String. This allowed the elimination of the clunky hashtable to relate
the Strings to the User objects. This required changing the toString of
User to return only the user name instead of a more complete
representation of User. Old functionality moved to fullString method.
(Issue #195)

Created an IDLE role to allow a dummy "Idle" user used as the bumper
between the active users and the idle users. Only the bumper/header
should use this role. The rendered automatically centers a "user" with
the IDLE role.

Fixed an issue where the lastActivity field was not being updated when a
user sent something to the server. The activity is now updated whenever
a message is received and the user list is broadcast to all.