Fixed UI issues with set show name and set host name.
Changed so text field popups for show name and show host are not updated
with updated text while they are open. This will prevent the strings
being updated while they are being edited. The text for the text field
is now updated when the popup is shown rather than as part of updateGUI.

The same was done for the text filter popup in AnswerQueuePanel, though
conflicting updates like this shouldn't be an issue.