Added top-level standings panel.
Added a top-level panel to display contest standings for any round. This
required a new class for the panel, updates to TriviaGUI and
AddTabDialog to account for the new available tab, and additional
settings in trivia-defaults. A new method, getLastAnnounced, was added
to Trivia to facilitate determining what the last announced round was.
Additional classes may be changed to use this method later. (Issue #175)

ScoreEntry was changed so that the default comparator is by score, then
by name. Capability to sort by team name was moved to new static method
alphabetize using a comparator that returns the comparison of team
names.

The team comparison chart was changed to use the new alphabetize method
to preserve functionality.