Added capability and GUI elements to track who is working on questions.
Trivia client now holds a User object instead of a String for user name
and a Role. Added a getUser method to return the user and removed
methods to get the user name or role, as they can now be accessed by
calling those methods on the user object. This also required downstream
changes wherever the user name was being retrieved by GUI elements from
the client.

Added field to User to hold the question number user is currently
working on (called effort). There is also a new method in Question that
takes a User[] in and returns all of the users working on that question
number.

Converted the question numbers on the open question panel into buttons
to indicate which question you are working on. This information is
passed to the server using a new client message type. A mouseover of the
button reveals who is working on that question. When a question is
closed, each user who was working on that question automatically has
their effort reset to 0, indicating no question being worked on. (Issue
#156)

Fixed bug where closing a question was not assigning the answer to the
correct question. Command was erroneously treating the qNumber as an
answer queue index.