Added indication of which visual trivia have been used/unused.
Added a new row to the bottom of the OpenQuestionsPanel that displays
all of the visual trivia numbers, with the numbers changing color if the
visual trivia is currently open or already used. This was done by
creating a new panel that spans the entire bottom and a series of labels
placed on that panel.

Added an int visualTrivia field to Question to hold the visual trivia
number (0 if none). Added a regex parse step to Question.setQuestion to
detect visual trivia and set visualTrivia. Added method
Question.getVisualTrivia to return the value of visualTrivia.

Added method to Trivia that iterates through every question and
determines whether each visual trivia has been used or not.

Removed abstract modified from TriviaPanel class as it has no abstract
methods.

Corrected bug in TriviaServer where it was broadcasting the number of
teams instead of the number of visual trivia.