Changed user list to display duration idle and in role instead of last.
Modified tooltip of users in UserList to display the time idle and time
since last role change instead of displaying the last date-time. This is
facilitated by new methods in user and a static method in UserListPanel,
durationToString to format the duration into a nice string.

Added methods to User to return a Duration between now and the last
active time or the last time the role changed. This also resulted in a
change from Date objects to LocalDateTime objects for lastActive and
lastRoleChange in order to use Duration. This in turn required the use
of Jackson JSR310 (added to dependencies) in order to
serialize/deserialize the LocalDateTime objects. Also created a JSON
constructor that takes in all of the fields, and changed existing
constructors to call that.

Fixed bug with User.compareTo method not properly sorting by Role before
by user name.

Fixed bug where server would change the user name when the role is set.
Since the user name is no longer set in the client message, this
resulted in the user's name being set to null.