Converted some Admin menu items to in-menu spinners.
Converted the Admin menu items for team number and number of visual
trivia to sub-menu spinners similar to the one for time-to-idle.

Gave all items that could trigger a ChangeEvent in TriviaFrame a name
and modified stateChange to determine the appropriate action based on
the name instead cascading ifs comparing the source to various
components.