Conversion to JRE7.
Previous attempts didn't actually compile with JRE7. The IntPredicate
method for checking array membership in OpenQuestionsPanel had to be
changed to use an IntBuffer and Integers. Sort for user lists in
UserListPanel had to be changed to indirect using Collections, as
ArrayList<T> does not have a sort method in JRE7. (Issue #198)