Fixed issue causing exceptions when a visual higher than max is used.
Changed getVisualTriviaUsed to do a bounds check on the visual trivia
specified by a question to make sure it is not greater than the number
of visual trivia. This was previously causing array out of bounds
exceptions that would render the OpenQuestionPanel unusable. (Issue
#204)

Changed the default number of visual trivia to 50.