Don't show role info for idle users; user list refresh timer
Changed the cell renderer so that role and time in role information is
only included in the tooltip if the user is not idle. (Issue #205)

Added a timer to UserListPanel to update the user list every five
seconds so idle users get moved to the idle section even if there are no
user list updates from the server. The timer calls updateGUI every 5s.