Created column to display the number of users working on a question.
Added a new column to OpenQuestionsPanel next to the question button
showing how many people are working on a question. Moved the mouseover
list of people working on the question to this new label.

Removed idleFrequency from server settings since it is unused now.

Removed space in team number from Summary Panel.