Fixed assorted small bugs during tech test.
Corrected a bug where the new round button was setting the round number
to the current round instead of the next round.

Corrected a bug where the new question dialog was using the wrong round
number during a speed round.

Corrected a bug where the team comparison chart would throw an exception
in certain situations due to an array out-of-bounds.

Partial correction to the incorrect path to visual trivia.