Ensure all clients receive updated standings after game load
Previously when loading games from a save file, the new game state would
not be transmitted to clients until after the standings had been loaded
from the KVSC website. Order was rearranged to transmit game data as
soon as it is loaded, then fetch and update standings and number of
teams.