Enable auto-round for history panel
The history panel now starts on the round previous to the current round
(except for round 1). The round displayed will also automatically update
to the previous round will automatically update when the round changes
as long as the round displayed had been the previous round before the
change.

Implements issue #246