Reorganize classes to separate instantiated panels and dialogs
The package was a bit unwieldy in trying to find specific classes, so
new subpackages were created to hold panels that are instantiated and
dialogs. Top-level structures and components remain in the client
package.

As part of this reorganization, the UserPanel was also moved inside the
SummaryPanel class, as that is the only panel that uses it. This
required some slight modifications to field and method visibility and/or
location.