Streamline Round class
A lot of the pass-through methods in Round were removed in favor of
accessing the Question or Answer instance. This reduces the amount of
code and hopefully makes the code easier to understand as well.

Methods in Round were also updated to use streams for efficiency.

Also introduced separate nQuestions field and corresponding setter to
allow for irregular numbers of questions.

Refs: #242