Check standing for current round
Server will now check for standings for the current round, instead of
just previous rounds. This is the easiest way to accommodate checking
for the final standings, rather than putting in a whole extra mechanism
for checking if the contest is over. The added cost from extra failed
fetches should be trivial.

Change-Id: Ic87966a346946ebe9db4531a894cd4f553306c8b