Allow a custom number of questions
There is now an option in the Admin menu to set a custom number of
questions for a round. To allow this change, the array of Questions in
Round was changed to an ArrayList<Question>. Round now starts with an
array of the normal amount of Questions, and whenever the number of
questions is changed, additional Questions are created. Questions are
*not* deleted if the number of questions is reduced in order to preserve
data for those questions in case of an accidental reduction in
questions. However, Round methods now only return Questions with
question numbers less than or equal to the current number of questions.

Changing the Question array to an ArrayList also required many
downstream changes in the Round methods.

A new message type was created to handle the new server interaction.

This change also required some changes to how question rows were made
visible/not visible in RoundQuestionsPanel.

SaveMediator was updated to read/write the number of questions for each
round. To preserve backwards compatibility with old saves, the number of
questions will be inferred from the speed status for saves where the
number of questions is missing.

A new field nQuestionsMax was added to Trivia to separate it from
nQuestionsSpeed. It is set in the server settings file to 36, which
should be more than enough.

Closes: #242

Change-Id: I6f554f0c9472d797a2a81776a4740e7644dbbbc4