Add search capability
Adds a new tab that allows users to search all questions and answers.
Query is processed as a regular expression, allowing complex searches.

Actual search functionality is processed within the Trivia class.
Results are returned as a LinkedHashMap sorted first by round number,
then by question number.

A modified version of the enterOverride capability was copied from
TriviaDialog to allow <Enter> to submit the search. In the future, these
may be unified.

The new search tab is added to the list of default tabs open when
starting.

Refs: #229

Change-Id: Ibe40683e16e62b20dfbad9357b6ccdfa8d08a30e