Fix NPE when no saves
Previously, if there were no save files when a user requested the save
list, the server would throw a NPE and the client would get no response.
Now if there are no save files, an empty string will be sent to the
client, and the server log will contain a message about no save files
being found, along with the save path.

Change-Id: I5ee94339cae1543139d52af0637c211a7994e469