Improve portability of audio capture script
The audio script now determines whether to use ffmpeg or avconv based on
which is available on the machine. It also uses the path of the script
to determine the destination. The script can now be used unaltered on
both trivia servers.

Also changed the break times to EST, as the time zones were being
ignored.

Change-Id: I7d3cb537b690d8cfd3e451bd3dac9da77b513d8d