Keep nQuestion spinner synchronized
The spinner in the admin menu to manually change the number of questions
in a round would not be updated if the number of questions in the round
changed because another user changed it or the round advanced. This
meant the starting value would not be the current number of questions in
a round. This value is now updated when the gui updates to match the
current number of questions.