Ensure nQuestions isn't reduced below latest question
Previously, users could adjust the number of questions below a question
that already had been opened, deleting all the data for that question.
Now the minimum bound is updated to the highest question number already
opened in the round.

Also removed some filters in Round that are no longer needed as the
question list only includes available questions.